+++ packages/index-page/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## Release
 
+### 1.0.3
+
+- Update donation-box-section.js: @twreporter/react-components/lib/link-with-tracker -> @twreporter/react-components/lib/donation-link-with-utm
+
 ### 1.0.2
 
 - Use `replaceGCSUrlOrigin` from `@twreporter/core` instead